MWPW-189894: Applying Nicolas's comment.
